rename components to better naming convention
